requirements.txt
Django==4.0.2
django-phonenumber-field==6.0.0
django-timezone-field==4.2.3
django-mptt==0.13.4
phonenumbers==8.12.42
Faker==12.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0